Modifies smart ping documentation
The paragraph "no on-demand host checks" is confusing and hides a detail about smart ping that should be immediately visible to the reader.

This commit restructures the "2. Smart Ping — Intelligent host checks" sections keeping the meaning intact, and moves in the description of smart ping's behaviour of listening for TCP packets and not UDP packets. The explanation of this behaviour is otherwise is hidden in the next paragraph, which could be easily overseen as thought as not connected to smart ping's description.