translate script: Fixed traceback in rare occasions
Sometimes several articles start with the same prefix additionally
to an overview article only containing that prefix. In such cases
the matching was too loose and all articles matching the prefix have
been captured. But it could happen that the originally searched article
is not available in a docs type while the others are. This lead to an
exception and is now fixed.